test: split nim tests per pages (#13)
