+++ frontend/src/__tests__/cypress/cypress/utils/nimUtils.ts
@@ -1,3 +1,34 @@
+import type { MockDashboardConfigType } from '~/__mocks__';
+import {
+  mock200Status,
+  mockDashboardConfig,
+  mockDscStatus,
+  mockK8sResourceList,
+  mockProjectK8sResource,
+  mockSecretK8sResource,
+} from '~/__mocks__';
+import {
+  AcceleratorProfileModel,
+  ConfigMapModel,
+  InferenceServiceModel,
+  ProjectModel,
+  PVCModel,
+  SecretModel,
+  ServingRuntimeModel,
+  TemplateModel,
+} from '~/__tests__/cypress/cypress/utils/models';
+import {
+  mockNimImages,
+  mockNimInferenceService,
+  mockNimModelPVC,
+  mockNimServingRuntime,
+  mockNimServingRuntimeTemplate,
+  mockNvidiaNimAccessSecret,
+  mockNvidiaNimImagePullSecret,
+} from '~/__mocks__/mockNimResource';
+import { mockAcceleratorProfile } from '~/__mocks__/mockAcceleratorProfile';
+import type { InferenceServiceKind } from '~/k8sTypes';
+
 export function findNimModelDeployButton(): Cypress.Chainable<JQuery> {
   return findNimModelServingPlatformCard().findByTestId('nim-serving-deploy-button');
 }
@@ -11,6 +42,7 @@ export function findServingPlatformLabel(): Cypress.Chainable<JQuery> {
 }
 
 export const modalDialogTitle = 'Deploy model with NVIDIA NIM';
+
 export function validateNvidiaNimModel(
   deployButtonElement: Cypress.Chainable<JQuery<HTMLElement>>,
 ): void {
@@ -123,3 +155,131 @@ export function validateNimInmferenceModelsTable(): void {
     'be.visible',
   );
 }
+
+/* ###################################################
+   ###### Interception Initialization Utilities ######
+   ################################################### */
+
+type EnableNimConfigType = {
+  hasAllModels?: boolean;
+};
+
+// intercept all APIs required for enabling NIM
+export const initInterceptsToEnableNim = ({ hasAllModels = false }: EnableNimConfigType): void => {
+  cy.interceptOdh(
+    'GET /api/dsc/status',
+    mockDscStatus({
+      installedComponents: {
+        kserve: true,
+        'model-mesh': true,
+      },
+    }),
+  );
+
+  cy.interceptOdh(
+    'GET /api/config',
+    mockDashboardConfig({
+      disableKServe: false,
+      disableModelMesh: false,
+      disableNIMModelServing: false,
+    }),
+  );
+
+  const project = mockProjectK8sResource({
+    hasAnnotations: true,
+    enableModelMesh: hasAllModels ? undefined : false,
+  });
+  if (project.metadata.annotations != null) {
+    project.metadata.annotations['opendatahub.io/nim-support'] = 'true';
+  }
+  cy.interceptK8sList(ProjectModel, mockK8sResourceList([project]));
+
+  const templateMock = mockNimServingRuntimeTemplate();
+  cy.interceptK8sList(TemplateModel, mockK8sResourceList([templateMock]));
+  cy.interceptK8s(TemplateModel, templateMock);
+
+  cy.interceptK8sList(
+    AcceleratorProfileModel,
+    mockK8sResourceList([mockAcceleratorProfile({ namespace: 'opendatahub' })]),
+  );
+
+  cy.intercept('GET', '/api/accelerators', {
+    configured: true,
+    available: { 'nvidia.com/gpu': 1 },
+    total: { 'nvidia.com/gpu': 1 },
+    allocated: { 'nvidia.com/gpu': 1 },
+  });
+};
+
+// intercept all APIs required for deploying new NIM models in existing projects
+export const initInterceptsToDeployModel = (nimInferenceService: InferenceServiceKind): void => {
+  cy.interceptK8s(ConfigMapModel, mockNimImages());
+  cy.interceptK8s('POST', SecretModel, mockSecretK8sResource({}));
+  cy.interceptK8s('POST', InferenceServiceModel, nimInferenceService).as('createInferenceService');
+
+  cy.interceptK8s('POST', ServingRuntimeModel, mockNimServingRuntime()).as('createServingRuntime');
+
+  // NOTES: `body` field is needed!
+  cy.intercept(
+    { method: 'GET', pathname: '/api/nim-serving/nvidia-nim-images-data' },
+    {
+      body: { body: mockNimImages() },
+    },
+  );
+  cy.intercept(
+    { method: 'GET', pathname: '/api/nim-serving/nvidia-nim-access' },
+    { body: { body: mockNvidiaNimAccessSecret() } },
+  );
+  cy.intercept('GET', 'api/nim-serving/nvidia-nim-image-pull', {
+    body: { body: mockNvidiaNimImagePullSecret() },
+  });
+  cy.interceptK8s('POST', PVCModel, mockNimModelPVC());
+};
+
+// intercept all APIs required for deleting an existing NIM models
+export const initInterceptsForDeleteModel = (): void => {
+  // create initial inference and runtime
+  cy.interceptK8sList(InferenceServiceModel, mockK8sResourceList([mockNimInferenceService()]));
+  cy.interceptK8sList(ServingRuntimeModel, mockK8sResourceList([mockNimServingRuntime()]));
+
+  // intercept delete inference request
+  cy.interceptK8s(
+    'DELETE',
+    {
+      model: InferenceServiceModel,
+      ns: 'test-project',
+      name: 'test-name',
+    },
+    mock200Status({}),
+  ).as('deleteInference');
+
+  // intercept delete runtime request
+  cy.interceptK8s(
+    'DELETE',
+    {
+      model: ServingRuntimeModel,
+      ns: 'test-project',
+      name: 'test-name',
+    },
+    mock200Status({}),
+  ).as('deleteRuntime');
+};
+
+// intercept all APIs required for verifying NIM enablement
+export const initInterceptorsValidatingNimEnablement = (
+  dashboardConfig: MockDashboardConfigType,
+  disableServingRuntime = false,
+): void => {
+  cy.interceptOdh('GET /api/config', mockDashboardConfig(dashboardConfig));
+
+  if (!disableServingRuntime) {
+    const templateMock = mockNimServingRuntimeTemplate();
+    cy.interceptK8sList(TemplateModel, mockK8sResourceList([templateMock]));
+    cy.interceptK8s(TemplateModel, templateMock);
+  }
+
+  cy.interceptK8sList(
+    ProjectModel,
+    mockK8sResourceList([mockProjectK8sResource({ hasAnnotations: true })]),
+  );
+};
